REF l10n_it_reverse_charge tests to be reused
